# public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: POS APis Documentation
  version: 1.0.0
  title: E Kang POS Frontend API's
basePath: /api/
paths:
...
  /category:
    get:
      tags:
        - Category
      summary: Manage Category
      description: 'Manage Category with this API'
      operationId: category
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
      security: 
        - api_key: []
    post:
      tags:
        - Category
      summary: Create Category
      description: 'Create Category with this API'
      operationId: createCategory
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Category Of Productions
          required: true
          schema:
            $ref: '#/definitions/Category' 
      responses:
        '200':
          description: successful operation
        '400':
          description: Bad request
      security:
        - api_key: [] 
...
securityDefinitions:
  api_key:
    type: apiKey
    name: Authorization
    in: header    
definitions:
...
 Category:
    type: object
    properties:
      name:
        type: string
        description: English Name
      kh_name:
        type: string
        description: Khmer Name
      description:
        type: string
        description: description of category
      brands:
        type: array
        items:
         type: integer 
        example: [1,2,3]  
    xml:
      name: Category
...